testing/geos: new aport
https://trac.osgeo.org/geos
The Geometry Engine Open Source (GEOS)
---
The OSGeo project maintains some very useful geospatial libraries
intended to benefit the community. Included in this project is a
low level geometry engine that is the under pinning of many open
source geospatial projects... GEOS. The library brings a number of
really cool spatial models and functions as well as encoders and
decoders for WKT and WKB. For those whom don't know, this library
contains  the really hard math associated with geospatial functionality.

This package brings the library to the x86_64 platform for Alpine and
has been tested on Alpine 3.8. Since we needed this library for an
internal project at WorldAware (and we're using Alpine as the basis
of several containers) it was easier to simply add a new package
descriptor rather than fold the library into our build process.

This also presented the opportunity for us to give back to the Alpine
community by sharing the package from the OSGeo project.

The package is released under the MIT license. We hope others find
this useful.